Split DoubleDownButton component
