+++ web/src/table_tennis/components/DoubleDownButton.js
@@ -1,0 +1,22 @@
+import { Box, Button } from "@chakra-ui/react";
+
+const DoubleDownButton = ({ dispatch }) => {
+  const handleDoubleDown = () => {
+    dispatch({ type: "double_down" });
+  };
+
+  return (
+    <Box py="4">
+      <Button
+        colorScheme="blue"
+        variant="outline"
+        mt="-2"
+        onClick={handleDoubleDown}
+      >
+        Double down
+      </Button>
+    </Box>
+  );
+};
+
+export default DoubleDownButton;
