+++ web/src/table_tennis/components/index.js
@@ -2,3 +2,4 @@ export { default as GameModeSwitch } from "./GameModeSwitch";
 export { default as ScoreSelector } from "./ScoreSelector";
 export { default as PlayerSelector } from "./PlayerSelector";
 export { default as LoadingPlaceholder } from "./LoadingPlaceholder";
+export { default as DoubleDownButton } from "./DoubleDownButton";
